pyproject.toml
[tool.poetry]
name = "test01"
version = "0.1.0"
description = ""
authors = ["sayedmehedi <sayedmehedi456@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
